調整檔案結構
