+++ app/enums/training_part.py
