+++ app/model/standard.py
@@ -1,7 +1,7 @@
 from mongoengine import Document, ListField, EnumField, IntField
 
-from enums.training_part import TrainingPart
-from enums.gender import Gender
+from app.enums.training_part import TrainingPart
+from app.enums.gender import Gender
 
 class Standard(Document):
     part = EnumField(TrainingPart)
