+++ requirements.txt
@@ -7,7 +7,7 @@ click==8.1.3
 colorama==0.4.5
 cryptography==37.0.2
 dnspython==2.2.1
-Flask==2.0.3
+Flask==2.1.3
 Flask-Cors==3.0.10
 Flask-Login==0.6.1
 Flask-Mail==0.9.1
